added custom prism factory
