+++ bundles/org.palladiosimulator.simexp.pcm.modelled.prism/META-INF/MANIFEST.MF
@@ -3,8 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: modelled prism
 Bundle-SymbolicName: org.palladiosimulator.simexp.pcm.modelled.prism
 Bundle-Version: 1.0.0.qualifier
-Export-Package: org.palladiosimulator.simexp.pcm.modelled.prism,
- org.palladiosimulator.simexp.pcm.modelled.prism.config
+Export-Package: org.palladiosimulator.simexp.pcm.modelled.prism
 Require-Bundle: org.apache.log4j;bundle-version="1.2.19",
  org.eclipse.emf.ecore;bundle-version="2.33.0",
  org.palladiosimulator.simexp.pcm;bundle-version="1.0.0",
