+++ bundles/org.palladiosimulator.simexp.pcm.modelled.prism/src/main/java/org/palladiosimulator/simexp/pcm/modelled/prism/ModelledPrismExperienceSimulationExecutorFactory.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 
 import org.palladiosimulator.experimentautomation.experiments.Experiment;
 import org.palladiosimulator.simexp.core.store.SimulatedExperienceStore;
+import org.palladiosimulator.simexp.pcm.config.IModelledPrismWorkflowConfiguration;
 import org.palladiosimulator.simexp.pcm.examples.deltaiot.provider.PrismMeasurementSpecificationProvider;
 import org.palladiosimulator.simexp.pcm.modelled.ModelledExperienceSimulationExecutorFactory;
 import org.palladiosimulator.simexp.pcm.modelled.ModelledModelLoader;
-import org.palladiosimulator.simexp.pcm.modelled.prism.config.IModelledPrismWorkflowConfiguration;
 import org.palladiosimulator.simexp.pcm.prism.entity.PrismSimulatedMeasurementSpec;
 import org.palladiosimulator.simulizar.reconfiguration.qvto.QVTOReconfigurator;
