--- bundles/org.palladiosimulator.simexp.pcm.modelled.prism/src/main/java/org/palladiosimulator/simexp/pcm/modelled/prism/config/IModelledPrismWorkflowConfiguration.java
@@ -1,8 +1,0 @@
-package org.palladiosimulator.simexp.pcm.modelled.prism.config;
-
-import org.palladiosimulator.simexp.pcm.config.IPrismConfiguration;
-import org.palladiosimulator.simexp.pcm.modelled.config.IModelledWorkflowConfiguration;
-
-public interface IModelledPrismWorkflowConfiguration extends IModelledWorkflowConfiguration, IPrismConfiguration {
-
-}
