+++ bundles/org.palladiosimulator.simexp.pcm.modelled.simulator/src/main/java/org/palladiosimulator/simexp/pcm/modelled/simulator/config/IModelledPcmWorkflowConfiguration.java
@@ -1,6 +1,6 @@
 package org.palladiosimulator.simexp.pcm.modelled.simulator.config;
 
-import org.palladiosimulator.simexp.pcm.modelled.config.IModelledWorkflowConfiguration;
+import org.palladiosimulator.simexp.pcm.config.IModelledWorkflowConfiguration;
 import org.palladiosimulator.simexp.pcm.simulator.config.IPCMConfiguration;
 
 public interface IModelledPcmWorkflowConfiguration extends IModelledWorkflowConfiguration, IPCMConfiguration {
