--- bundles/org.palladiosimulator.simexp.pcm.modelled/src/main/java/org/palladiosimulator/simexp/pcm/modelled/config/IModelledWorkflowConfiguration.java
@@ -1,7 +1,0 @@
-package org.palladiosimulator.simexp.pcm.modelled.config;
-
-import org.palladiosimulator.simexp.pcm.config.IWorkflowConfiguration;
-
-public interface IModelledWorkflowConfiguration extends IWorkflowConfiguration, IModelledConfiguration {
-
-}
