+++ bundles/org.palladiosimulator.simexp.pcm/src/org/palladiosimulator/simexp/pcm/config/IModelledConfiguration.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.simexp.pcm.modelled.config;
+package org.palladiosimulator.simexp.pcm.config;
 
 import org.eclipse.emf.common.util.URI;
