+++ bundles/org.palladiosimulator.simexp.pcm/src/org/palladiosimulator/simexp/pcm/config/IModelledPrismWorkflowConfiguration.java
@@ -1,0 +1,5 @@
+package org.palladiosimulator.simexp.pcm.config;
+
+public interface IModelledPrismWorkflowConfiguration extends IModelledWorkflowConfiguration, IPrismConfiguration {
+
+}
