+++ bundles/org.palladiosimulator.simexp.pcm/src/org/palladiosimulator/simexp/pcm/config/IModelledWorkflowConfiguration.java
@@ -1,0 +1,5 @@
+package org.palladiosimulator.simexp.pcm.config;
+
+public interface IModelledWorkflowConfiguration extends IWorkflowConfiguration, IModelledConfiguration {
+
+}
