+++ bundles/org.palladiosimulator.simexp.workflow/META-INF/MANIFEST.MF
@@ -16,27 +16,16 @@ Require-Bundle: org.palladiosimulator.analyzer.workflow,
  org.palladiosimulator.simexp.core.store.csv,
  org.palladiosimulator.simexp.pcm,
  tools.mdsd.probdist.api;bundle-version="1.0.0",
- tools.mdsd.probdist.api.apache,
- tools.mdsd.probdist.model.basic,
- tools.mdsd.probdist.model,
- org.palladiosimulator.simexp.pcm.examples.deltaiot,
  org.palladiosimulator.simexp.pcm,
  org.palladiosimulator.experimentautomation,
- org.palladiosimulator.envdyn.api,
  org.palladiosimulator.simexp.pcm.examples,
- org.palladiosimulator.simexp.pcm.prism,
  org.apache.commons.lang;bundle-version="2.6.0",
  org.palladiosimulator.simexp.markovian.model,
  org.palladiosimulator.simulizar.reconfiguration.qvto,
- org.palladiosimulator.simexp.pcm.reliability,
- org.palladiosimulator.solver,
  org.palladiosimulator.dependability.reliability.uncertainty,
  org.palladiosimulator.simexp.pcm.examples.deltaiot.param,
- org.palladiosimulator.simexp.pcm.performance,
- org.palladiosimulator.simexp.pcm.performability,
  org.palladiosimulator.simexp.pcm.simulator;bundle-version="1.0.0",
  org.palladiosimulator.simexp.pcm.modelled;bundle-version="1.0.0",
- org.palladiosimulator.simexp.pcm.modelled.prism;bundle-version="1.0.0",
  org.palladiosimulator.simexp.pcm.modelled.simulator;bundle-version="1.0.0",
  org.palladiosimulator.simexp.model.io;bundle-version="1.0.0",
  org.palladiosimulator.simexp.dsl.smodel;bundle-version="0.1.0",
