+++ bundles/org.palladiosimulator.simexp.workflow/src/main/java/org/palladiosimulator/simexp/workflow/launcher/PrismSimulationExecutorFactory.java
@@ -3,12 +3,8 @@ package org.palladiosimulator.simexp.workflow.launcher;
 import java.util.Optional;
 
 import org.palladiosimulator.core.simulation.SimulationExecutor;
-import org.palladiosimulator.simexp.core.entity.SimulatedMeasurementSpecification;
 import org.palladiosimulator.simexp.core.store.DescriptionProvider;
-import org.palladiosimulator.simexp.core.store.SimulatedExperienceStore;
 import org.palladiosimulator.simexp.pcm.config.IPrismWorkflowConfiguration;
-import org.palladiosimulator.simexp.pcm.examples.deltaiot.DeltaIoTSimulationExecutorFactory;
-import org.palladiosimulator.simexp.pcm.examples.executor.PcmExperienceSimulationExecutorFactory;
 
 import tools.mdsd.probdist.api.random.ISeedProvider;
 
@@ -16,10 +12,11 @@ public class PrismSimulationExecutorFactory extends BaseSimulationExecutorFactor
 
     public SimulationExecutor create(IPrismWorkflowConfiguration workflowConfiguration,
             DescriptionProvider descriptionProvider, Optional<ISeedProvider> seedProvider) {
-        PcmModelLoader.Factory modelLoaderFactory = new PcmModelLoader.Factory();
-        PcmExperienceSimulationExecutorFactory<? extends Number, ?, ? extends SimulatedMeasurementSpecification> factory = new DeltaIoTSimulationExecutorFactory(
-                workflowConfiguration, modelLoaderFactory, new SimulatedExperienceStore<>(descriptionProvider),
-                seedProvider);
-        return factory.create();
+//        PcmModelLoader.Factory modelLoaderFactory = new PcmModelLoader.Factory();
+//        PcmExperienceSimulationExecutorFactory<? extends Number, ?, ? extends SimulatedMeasurementSpecification> factory = new DeltaIoTSimulationExecutorFactory(
+//                workflowConfiguration, modelLoaderFactory, new SimulatedExperienceStore<>(descriptionProvider),
+//                seedProvider);
+//        return factory.create();
+        return null;
     }
 }
