+++ examples/org.palladiosimulator.simexp.pcm.examples.deltaiot/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Deltaiot
-Bundle-SymbolicName: org.palladiosimulator.simexp.pcm.examples.deltaiot
+Bundle-SymbolicName: org.palladiosimulator.simexp.pcm.examples.deltaiot;singleton:=true
 Bundle-Version: 1.0.0.qualifier
 Export-Package: org.palladiosimulator.simexp.pcm.examples.deltaiot,
  org.palladiosimulator.simexp.pcm.examples.deltaiot.process,
@@ -35,4 +35,6 @@ Require-Bundle: org.apache.log4j;bundle-version="1.2.15",
  org.palladiosimulator.simulizar.reconfiguration.qvto;bundle-version="4.3.0",
  de.uka.ipd.sdq.workflow.mdsd;bundle-version="4.3.0",
  de.uka.ipd.sdq.scheduler;bundle-version="5.2.0",
- org.apache.commons.csv;bundle-version="1.8.0"
+ org.apache.commons.csv;bundle-version="1.8.0",
+ org.palladiosimulator.simexp.commons.constants,
+ org.palladiosimulator.simexp.workflow;bundle-version="1.0.0"
