+++ examples/org.palladiosimulator.simexp.pcm.examples.deltaiot/build.properties
@@ -1,4 +1,5 @@
 source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
-               .
+               .,\
+               plugin.xml
