+++ examples/org.palladiosimulator.simexp.pcm.examples.deltaiot/plugin.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.palladiosimulator.simexp.workflow.launch">
+      <factory
+            class="org.palladiosimulator.simexp.pcm.examples.deltaiot.PrismSimulationExecutorLaunchFactory">
+      </factory>
+   </extension>
+
+</plugin>
