+++ features/org.palladiosimulator.simexp.feature/feature.xml
@@ -153,4 +153,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="org.palladiosimulator.envdyn.examples.deltaiot"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
